+++ packages/react/src/components/listbox/index.ts
@@ -1,8 +1,9 @@
 export type {
-  HighlightChangeDetails,
-  ScrollToIndexDetails,
-  SelectionDetails,
-  ValueChangeDetails,
+  HighlightChangeDetails as ListboxHighlightChangeDetails,
+  ScrollToIndexDetails as ListboxScrollToIndexDetails,
+  SelectionDetails as ListboxSelectionDetails,
+  SelectionMode as ListboxSelectionMode,
+  ValueChangeDetails as ListboxValueChangeDetails,
 } from '@zag-js/listbox'
 export { createListCollection, type CollectionItem, type ListCollection, type SelectionMode } from '../collection'
 export { ListboxContent, type ListboxContentBaseProps, type ListboxContentProps } from './listbox-content'