+++ packages/vue/src/components/combobox/examples/advanced.vue
@@ -3,11 +3,13 @@ import { Combobox, createListCollection } from '@ark-ui/vue/combobox'
 
 const frameworks = createListCollection({
   items: [
-    { label: 'React', value: 'react' },
-    { label: 'Solid', value: 'solid' },
-    { label: 'Vue', value: 'vue' },
-    { label: 'Svelte', value: 'svelte', disabled: true },
+    { label: 'React', value: 'react', type: 'JS' },
+    { label: 'Solid', value: 'solid', type: 'JS' },
+    { label: 'Vue', value: 'vue', type: 'JS' },
+    { label: 'Panda', value: 'panda', type: 'CSS' },
+    { label: 'Tailwind', value: 'tailwind', type: 'CSS' },
   ],
+  groupBy: (item) => item.type,
 })
 </script>
 
@@ -22,9 +24,9 @@ const frameworks = createListCollection({
     <Teleport to="body">
       <Combobox.Positioner>
         <Combobox.Content>
-          <Combobox.ItemGroup>
-            <Combobox.ItemGroupLabel>Frameworks</Combobox.ItemGroupLabel>
-            <Combobox.Item v-for="item in frameworks.items" :key="item.value" :item="item">
+          <Combobox.ItemGroup :key="type" v-for="[type, group] in frameworks.group()">
+            <Combobox.ItemGroupLabel>{{ type }}</Combobox.ItemGroupLabel>
+            <Combobox.Item v-for="item in group" :key="item.value" :item="item">
               <Combobox.ItemText>{{ item.label }}</Combobox.ItemText>
               <Combobox.ItemIndicator>✓</Combobox.ItemIndicator>
             </Combobox.Item>