+++ packages/vue/src/components/listbox/examples/group.vue
@@ -3,11 +3,13 @@ import { Listbox, createListCollection } from '@ark-ui/vue/listbox'
 
 const collection = createListCollection({
   items: [
-    { label: 'React', value: 'react' },
-    { label: 'Solid', value: 'solid' },
-    { label: 'Svelte', value: 'svelte', disabled: true },
-    { label: 'Vue', value: 'vue' },
+    { label: 'React', value: 'react', type: 'JS' },
+    { label: 'Solid', value: 'solid', type: 'JS' },
+    { label: 'Vue', value: 'vue', type: 'JS' },
+    { label: 'Panda', value: 'panda', type: 'CSS' },
+    { label: 'Tailwind', value: 'tailwind', type: 'CSS' },
   ],
+  groupBy: (item) => item.type,
 })
 </script>
 
@@ -15,9 +17,9 @@ const collection = createListCollection({
   <Listbox.Root :collection="collection">
     <Listbox.Label>Select your Frameworks</Listbox.Label>
     <Listbox.Content>
-      <Listbox.ItemGroup>
-        <Listbox.ItemGroupLabel>JS Frameworks</Listbox.ItemGroupLabel>
-        <Listbox.Item v-for="item in collection.items" :key="item.value" :item="item">
+      <Listbox.ItemGroup v-for="[type, group] in collection.group()" :key="type">
+        <Listbox.ItemGroupLabel>{{ type }}</Listbox.ItemGroupLabel>
+        <Listbox.Item v-for="item in group" :key="item.value" :item="item">
           <Listbox.ItemText>{{ item.label }}</Listbox.ItemText>
           <Listbox.ItemIndicator />
         </Listbox.Item>